back/requirements.txt
Django==4.2.20
asgiref==3.8.1
sqlparse==0.5.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
dj-database-url==2.3.0
python-dotenv==1.1.0
python-dotenv==1.1.0
dj-database-url